requirements.txt
streamlit==1.36.0
spotipy==2.23.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
python-dotenv==1.0.1